+++ SpotNetCore/SpotNetCore/Implementation/CommandHandler.cs
@@ -2,7 +2,6 @@ using System;
 using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
-using System.Text;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using SpotNetCore.Models;
@@ -473,7 +472,7 @@ namespace SpotNetCore.Implementation
             
             var split = input.Split(new[] { "-", "--" }, StringSplitOptions.RemoveEmptyEntries);
             //var split = input.Split(" ");
-            var p =new ParsedCommand
+            return new ParsedCommand
             {
                 Command = split[0].Trim(),
                 Parameters = split.Skip(1).Select(x =>
@@ -486,65 +485,6 @@ namespace SpotNetCore.Implementation
                     };
                 })
             };
-
-            return p;
-        }
-
-        public class ConsoleInput
-        {
-            private StringBuilder Buffer { get; } = new StringBuilder();
-            public int CurrentIndex { get; set; }
-
-            public ConsoleInput Append(String str)
-            {
-                Buffer.Append(str);
-                CurrentIndex += str.Length;
-                return this;
-            }
-            public ConsoleInput Append(Char c)
-            {
-                Buffer.Append(c);
-                CurrentIndex++;
-                return this;
-            }
-
-            public ConsoleInput Insert(char c, int index = -1) => Insert(c.ToString(), index);
-
-            public ConsoleInput Insert(String str, int index = -1)
-            {
-                if (index == -1)
-                {
-                    index = CurrentIndex;
-                }
-                Buffer.Insert(index, str);
-                CurrentIndex = index + str.Length;
-                return this;
-            }
-
-            public ConsoleInput Clear()
-            {
-                Buffer.Clear();
-                CurrentIndex = 0; 
-                return this;
-            }
-
-            public ConsoleInput Remove(int offset, int count)
-            {
-                Buffer.Remove(offset, count);
-                CurrentIndex = offset - count + 1;
-                return this;
-            }
-
-            public override string ToString()
-            {
-                return Buffer.ToString();
-            }
-
-            public static implicit operator string(ConsoleInput consoleInput)
-            {
-                return consoleInput.ToString();}
-
-            public int Length => Buffer.Length;
         }
     }
 }
